requirements.txt
# Core ADK dependencies
google-adk>=1.0.0
google-cloud-aiplatform[adk]>=1.93.0

# Google Cloud services
google-cloud-storage>=2.10.0
google-auth>=2.20.0

# Vertex AI RAG Engine
vertexai>=1.93.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data processing
requests>=2.31.0

# Optional: Enhanced vector search capabilities
# google-cloud-discoveryengine>=0.11.0  # For Vertex AI Search integration
# weaviate-client>=4.0.0                # For Weaviate vector database
# pinecone-client>=3.0.0                # For Pinecone vector database

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

google-genai>=1.9.0
google-cloud-firestore>=2.16.0
langchain>=0.1.0
langchain-google-vertexai>=1.0.0
langchain-community>=0.0.20
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
httpx>=0.26.0
pypdf2>=3.0.1
python-multipart>=0.0.6
markdown>=3.5.2
black>=25.1.0
jupyter>=1.0.0 